+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed docstrings with missing Notes sections causing spurious parameters in Sphinx documentation.
 - Fixed coordinate snapping in shape gradient computation to handle 0 normal and grid spacing components.
 - Fixed duplicate printing of cost estimation of `Job`.
+- Fixed EME with inconsistent `interp_spec`.
 
 ## [2.10.1] - 2026-01-14